Clarification needed on turning on RI Sharing on Member Account
This article says: 
"To share an RI with an account, both accounts must have sharing turned on"


I have a payer and a member account, but I don't see the option to turn on RI Sharing on the member account. I only see this message: "Linked account preferences are managed at the payer account level." It seems to me that you can only control turn on or off the RI sharing to the payer account only. 

I haven't purchased any Reserved Instances on the member account so perhaps, I might be able to see and edit the RI Sharing settings on the member account once I have done this. Yet still, the statement is misleading since you can share your RI by setting it on the payer account alone, without any configuration on the member account. Could you kindly confirm that you really can turn on RI Sharing on the member account?