Incorrect description
Incorrect description that Activate IAM access setting is always disabled. It is not. If you create an account with AWS Organizations API and do not specify the parameter, it is set to ALLOW (enabled). I think removing this ambiguous statement is useful plus it does not take away any information fidelity shared on this page. But having this wrong description may lead to sec issues